requirements.txt
pathlib2==2.3.7.post1
Pillow==10.0.0
Pillow==10.0.0
pytest==7.4.0
rasterio==1.3.8
setuptools==65.5.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
vit_pytorch==1.2.4
